+++ CHANGELOG.md
@@ -2,6 +2,7 @@ PowerModels.jl Change Log
 =========================
 
 ### Staged
+- Added SOC formulation of the storage model
 - Simplified PowerModel types (#192) (breaking)
 - Improved multi-conductor support in the storage model (#572) (breaking)
 - Renamed model_constructor to model_type (#192) (breaking)